+++ src/Models/Operations/GetSessionListRequest.php
@@ -40,10 +40,10 @@ class GetSessionListRequest
      *
      * Can be used for paginating the results together with `offset`.
      *
-     * @var ?float $limit
+     * @var ?int $limit
      */
     #[SpeakeasyMetadata('queryParam:style=form,explode=true,name=limit')]
-    public ?float $limit = null;
+    public ?int $limit = null;
 
     /**
      * Skip the first `offset` results when paginating.
@@ -51,19 +51,19 @@ class GetSessionListRequest
      * Needs to be an integer greater or equal to zero.
      * To be used in conjunction with `limit`.
      *
-     * @var ?float $offset
+     * @var ?int $offset
      */
     #[SpeakeasyMetadata('queryParam:style=form,explode=true,name=offset')]
-    public ?float $offset = null;
+    public ?int $offset = null;
 
     /**
      * @param  ?string  $clientId
      * @param  ?string  $userId
      * @param  ?Status  $status
-     * @param  ?float  $limit
-     * @param  ?float  $offset
+     * @param  ?int  $limit
+     * @param  ?int  $offset
      */
-    public function __construct(?string $clientId = null, ?string $userId = null, ?Status $status = null, ?float $limit = 10, ?float $offset = 0)
+    public function __construct(?string $clientId = null, ?string $userId = null, ?Status $status = null, ?int $limit = 10, ?int $offset = 0)
     {
         $this->clientId = $clientId;
         $this->userId = $userId;
